[Feat] LockScreen LiveActivity 첫동작 구현
