+++ Project_Timer/Global/sharedClass/DailyManager.swift
