+++ Project_Timer/Global/sharedClass/MotionDetector.swift
